fix main centralization bug
Main content was not centralized in small devices, so I fixed it by making two things:

First, I made the .articles class to 80% to guarantee that the main content will always be 80% of the page, and then I used justify-content on the articles to keep them on the center. The wrap was making the amount of space in the container bigger so it wouldn't centralize, however, with this settign, this is no longer an issue, because it is not only centralizing the container .articles, but it is as well centralizing the elements inside articles.